Install script for java
There are Java8 specific things that don't work in Java 9. Things like Lombok, which are super important to the projects right now. We need to find a viable way to upgrade them, but for the moment that will take time that I don't have. Let's create a shellscript to make our lives easier for the moment.